Move PromptsList classes to bloggingprompts.list package
