+++ WordPress/src/main/java/org/wordpress/android/ui/bloggingprompts/list/BloggingPromptsListParentViewModel.kt
@@ -1,17 +1,17 @@
-package org.wordpress.android.ui.bloggingprompts
+package org.wordpress.android.ui.bloggingprompts.list
 
 import androidx.annotation.StringRes
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.wordpress.android.R
 import org.wordpress.android.fluxc.model.SiteModel
-import org.wordpress.android.ui.bloggingprompts.PromptSection.ALL
-import org.wordpress.android.ui.bloggingprompts.PromptSection.ANSWERED
-import org.wordpress.android.ui.bloggingprompts.PromptSection.NOT_ANSWERED
+import org.wordpress.android.ui.bloggingprompts.list.PromptSection.ALL
+import org.wordpress.android.ui.bloggingprompts.list.PromptSection.ANSWERED
+import org.wordpress.android.ui.bloggingprompts.list.PromptSection.NOT_ANSWERED
 import javax.inject.Inject
 
 @HiltViewModel
-class BloggingPromptsListViewModel @Inject constructor(
+class BloggingPromptsListParentViewModel @Inject constructor(
     private val provider: BloggingPromptsListSiteProvider,
     private val analyticsTracker: BloggingPromptsListAnalyticsTracker,
 ) : ViewModel() {
