+++ WordPress/src/main/res/layout/blogging_prompts_activity.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/fragment_container"
-        android:name="org.wordpress.android.ui.bloggingprompts.BloggingPromptsListParentFragment"
+        android:name="org.wordpress.android.ui.bloggingprompts.list.BloggingPromptsListParentFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
