+++ WordPress/src/test/java/org/wordpress/android/ui/bloggingprompts/BloggingPromptsListAnalyticsTrackerTest.kt
@@ -7,6 +7,9 @@ import org.wordpress.android.BaseUnitTest
 import org.wordpress.android.analytics.AnalyticsTracker.Stat.BLOGGING_PROMPTS_LIST_ACCESSED
 import org.wordpress.android.analytics.AnalyticsTracker.Stat.BLOGGING_PROMPTS_LIST_TAB_CHANGED
 import org.wordpress.android.fluxc.model.SiteModel
+import org.wordpress.android.ui.bloggingprompts.list.BloggingPromptsListAnalyticsTracker
+import org.wordpress.android.ui.bloggingprompts.list.PromptSection
+import org.wordpress.android.ui.bloggingprompts.list.TRACKS_SELECTED_TAB
 import org.wordpress.android.util.analytics.AnalyticsTrackerWrapper
 
 class BloggingPromptsListAnalyticsTrackerTest : BaseUnitTest() {
